code/ch10/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
deprecation==2.1.0
distro==1.9.0
dynaconf==3.2.12
evidently==0.7.20
Faker==40.4.0
filelock==3.20.3
fsspec==2026.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
iterative-telemetry==0.0.10
joblib==1.5.3
litestar==2.19.0
litestar-htmx==0.5.0
markdown-it-py==4.0.0
mdurl==0.1.2
msgspec==0.20.0
multidict==6.7.1
multipart==1.3.0
mypy_extensions==1.1.0
nltk==3.9.2
numpy==2.4.2
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pandas==3.0.0
patsy==1.0.2
plotly==5.24.1
polyfactory==3.2.0
protobuf==6.33.5
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
rich-click==1.9.7
scikit-learn==1.8.0
scipy==1.17.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
statsmodels==0.14.6
tenacity==9.1.4
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.40.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
zipp==3.23.0
